requirements.txt
Django>=6.0,<7.0
# Lock to 3.0.1 until https://github.com/farridav/django-jazzmin/issues/651 is resolved
django-jazzmin==3.0.1
django-import-export
psycopg2-binary
django-storages[boto3]
django-ses[events]
django-post-office
django-logentry-admin
gunicorn
whitenoise
Markdown
Pillow
python-freeipa==1.0.10
django-avatar
pyotp
qrcode
pycountry
python-dateutil
scikit-learn>=1.3.0
valx
sentry-sdk[django]
ruff
